requirements.txt
kafka-python
psycopg2-binary
fastapi
uvicorn
python-dotenv
requests
hypothesis
clickhouse-driver
